feat: added dao for catalog models & moved all services in CatalogService & ImporterService
